+++ finstack-quant/statements-analytics/tests/checks/mod.rs
@@ -2,7 +2,6 @@
 //!
 mod consistency_tests;
 mod credit_tests;
-mod formula_check_tests;
 mod golden_tests;
 mod integration_tests;
 mod reconciliation_tests;