+++ finstack-quant/statements/tests/checks/mod.rs
@@ -3,6 +3,7 @@
 mod balance_sheet_tests;
 mod cash_reconciliation_tests;
 mod data_quality_tests;
+mod formula_check_tests;
 mod proptest_checks;
 mod retained_earnings_tests;
 mod runner_tests;